==> week3/10-logistic_regression_ex.ipynb <==
from sklearn import datasets
x,y=datasets.fetch_openml("mnist_784", version=1, return_X_y=True)

x.head()

x.shape      

y.head() 

y.shape     

import matplotlib.pyplot as plt
import numpy as np

fig,axes=plt.subplots(nrows=3, ncols=3, figsize=(3,3))
for i,ax in enumerate(axes.flat):   #i -> index değerleri ax-> pixels
    digit=x.iloc[i]
    digit_pixels=np.array(digit).reshape(28,28)
    ax.imshow(digit_pixels)
    ax.set_title(f"örnek {i+1}")
plt.tight_layout()
plt.show()

import matplotlib.pyplot as plt
import numpy as np

fig,axes=plt.subplots(nrows=3, ncols=3, figsize=(6,6))
for i,ax in enumerate(axes.flat):   #i -> index değerleri ax-> pixels
    digit=x.iloc[i]
    digit_pixels=np.array(digit).reshape(28,28)
    ax.imshow(digit_pixels)
    ax.set_title(f"örnek {i+1}")
plt.tight_layout()
plt.show()

from sklearn.model_selection import train_test_split
xtrain, xtest, ytrain, ytest = train_test_split(x,y,test_size=0.2,random_state=42, stratify=True)  

from sklearn.linear_model import LogisticRegression
lr=LogisticRegression(max_iter=1000)
lr.fit(xtrain,ytrain)

from sklearn.metrics import accuracy_score
ypred=lr.predict(xtest)
accuracy=accuracy_score(ypred,ytest)
print("Accuracy", accuracy)

from sklearn.metrics import confusion_matrix
import seaborn as sns
cm=confusion_matrix(ytest, ypred)
fig,ax=plt.subplots(figsize=(8,6))
sns.heatmap(cm,annot=True,cmap="Blues",ax=ax,fmt="d",annot_kws={"size":12})
ax.set_xlabel("prediction", fontsize=14)
ax.set_ylabel("truth", fontsize=16)

from PIL import Image
import numpy as numpy
import matplotlib.pyplot as plt

img=Image.open("./test-mnist/6.png")  

img=Image.open("6.png")
img=img.resize((28,28))
img=img.convert("L")
img_array=np.array(img).reshape(1,-1)

pred=lr.predict(img_array)

pred

#img_array=np.array(img).flatten()

print(f"Predicted number {pred}")
plt.imshow(img)

#save the model
import pickle

with open("model.pkl","wb") as file:      #write binary
    pickle.dump(lr,file)

with open("model.pkl", "rb") as file:     #read binary
    loaded_model=pickle.looad(file)

We need to save the scaler separately because if we want to use the scaler later and access normal data, we need to call it too!!!

loaded_model.predict(img_array)
