==> week3/12-ex_with_exportedmodel.ipynb <==
import pickle
with open("scalar.pkl","rb") as file:
    loadeed_scaler=pickle.load(file)

with open("model.pkl","rb") as file:
    loaded_model=pickle.load(file)

import numpy as np
sample_data = [17.99,10.38,122.8,1001,0.1184,0.2776,0.3001,0.1471,0.2419,0.07871,1.095,0.9053,8.589,153.4,0.006399,0.04904,0.05373,0.01587,0.03003,0.006193,25.38,17.33,184.6,2019,0.1622,0.6656,0.7119,0.2654,0.4601,0.1189]

import pandas as pd
manuel_df=pd.DataFrame([sample_data])
manuel_df

manuel_scalar=loadeed_scaler.transform(manuel_df)

manuel_scalar

prediction = loaded_model.predict(manuel_scalar)

prediction

print(f"prediciton: {'malign' if prediction[0]==1 else 'benign'}")
