==> week3/14-titanic_classification_ex.ipynb <==
import numpy as np
import pandas as pd

data=pd.read_csv("titanic.csv")

data.head()

data.drop(["PassengerId", "Name", "Ticket", "Cabin"], axis=1, inplace=True)   #remove non-usable data

data.head()

from sklearn.preprocessing import LabelEncoder
lb = LabelEncoder()
data.Sex=lb.fit_transform(data.Sex)     #convert sex type to a number 0/1

data

df=pd.DataFrame(data)

df

df["Age"].fillna((df["Age"]).mean(),inplace=True)    #fill the missing values

# data1=pd.get_dummies(data, columns=["Embarked"])

data.Embarked=lb.fit_transform(data.Embarked) 

data

y=data.Survived.values
X=data.drop(["Survived"], axis=1)

from sklearn.model_selection import train_test_split
Xtrain, Xtest, ytrain, ytest = train_test_split(X,y,test_size=0.2, random_state=42)

from sklearn.preprocessing import StandardScaler
sc=StandardScaler()

Xtrain=sc.fit_transform(Xtrain)  
Xtest=sc.transform(Xtest)   

from sklearn.tree import DecisionTreeClassifier
dc=DecisionTreeClassifier(max_depth=3)

dc.fit(Xtrain,ytrain)

yhead=dc.predict(Xtest)

dc.score(Xtest, ytest)

# sample_data = np.array([[3, 0, 22, 1, 0, 7.2500, 0]])
sample_data = np.array([[1,0,19.000000,0,0,30.0000,2]])

scaler = StandardScaler()
sample_data = scaler.fit_transform(sample_data).reshape(1,-1)
prediction = dc.predict(sample_data)

if prediction == 1:
    print("Survived")
else:
    print("Non-Survived")
