==> week3/15-diabetes_classification_ex.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data=pd.read_csv("diabetes.csv")

data.head()

y=data.Outcome.values
X=data.drop(["Outcome"],axis=1)

from sklearn.model_selection import train_test_split
Xtrain,Xtest,ytrain,ytest = train_test_split(X,y,test_size=0.2,random_state=42)

from sklearn.preprocessing import StandardScaler
sc=StandardScaler()

Xtrain=sc.fit_transform(Xtrain)
Xtest=sc.transform(Xtest)

from sklearn.neighbors import KNeighborsClassifier
knn=KNeighborsClassifier(n_neighbors=4)

knn.fit(Xtrain,ytrain)

yhead=knn.predict(Xtest)

knn.score(Xtest,ytest)

yhead

new_predict=knn.predict(sc.transform(np.array([[6,148,72,35,0,33.6,0.627,50]])))
new_predict

import pickle
with open("scaler.pkl","wb") as file:
    pickle.dump(sc,file)

with open("model.pkl","wb") as file:
    pickle.dump(knn,file)
