==> week3/16-flask_ex.ipynb <==
import pickle
with open("scaler.pkl","rb") as file:
    load_sc=pickle.load(file)

with open("model.pkl","rb") as file:
    load_model=pickle.load(file)

import numpy as np
new_predict=load_model.predict(load_sc.transform(np.array([[6,148,72,35,0,33.6,0.627,50]])))
print(new_predict)

from flask import Flask, request, render_template

app=Flask(__name__)
@app.route("/",methods=["GET", "POST"])
def get_predict():
    prediction=None     
    prediction1=None     
    if request.method=="POST":
        Pregnancies=float(request.form["Pregnancies"])
        Glucose=float(request.form["Glucose"])
        BloodPressure=float(request.form["BloodPressure"])
        SkinThickness=float(request.form["SkinThickness"])
        Insulin=float(request.form["Insulin"])
        BMI=float(request.form["BMI"])
        DiabetesPedigreeFunction=float(request.form["DiabetesPedigreeFunction"])
        Age=float(request.form["Age"])
        
        
        user_info=np.array([[Pregnancies, Glucose,BloodPressure, SkinThickness, Insulin, BMI,  DiabetesPedigreeFunction, Age ]])
        prediction1=load_model.predict(load_sc.transform(user_info))
    return render_template("web.html", prediction=prediction1)


if __name__== "__main__":
    app.run(port=5001)

#by "ngrok" you can share on web site this project
#provides directing the port 

# ngrok http port 5001 terminale yaz
