==> week3/18-classifier_ex.ipynb <==
import numpy as np
import pandas as pd
import os
import PIL.Image as img

bacterial_leaf_blight="rice_leaf_diseases/Bacterial leaf blight"
brown_spot="rice_leaf_diseases/Brown spot"
leaf_smut="rice_leaf_diseases/Leaf smut"
test="rice_leaf_diseases/test/test.png"

def file(path):
    return [os.path.join(path,f) for f in os.listdir(path)]    # enters the folder from the path entered as a parameter, reads all folders inside and returns the name of each folder

def convert_data(file_name, class_name):
    photos=file(file_name)
    photo_class=[]
    
    for photo in photos:   
        read_photo = img.open(photo).convert("L")    #grayscale -> decreasing efficiency on colour images
        resize_photo = read_photo.resize((28,28))  #for second bracket is in tuple format, resize the photo by 28x28
        normalize_photo = np.array(resize_photo)/255.0  #There are 255 values in total, since the photo is coloured, we actually scale the photo between 0-255 -> increases efficiency and learning
        convert_photo=normalize_photo.flatten()  
        
        if class_name == "bacterial_leaf_blight":
            data = np.append(convert_photo, [0])  
        elif class_name == "brown_spot":
            data = np.append(convert_photo, [1])   
        elif class_name == "leaf_smut":
            data = np.append(convert_photo, [2])
        else:
            continue
        
        photo_class.append(data)
    return photo_class
        

data_bacterial_leaf_blight = convert_data(bacterial_leaf_blight, "bacterial_leaf_blight")    #We created two dataframes named covid and non-covid and in the last column it returns its class, if he/she is covid it writes 1, is not 0
bacterial_leaf_blight_df = pd.DataFrame(data_bacterial_leaf_blight)
data_brown_spot = convert_data(brown_spot, "brown_spot")
brown_spot_df = pd.DataFrame(data_brown_spot)
data_leaf_smut = convert_data(leaf_smut, "leaf_smut")
leaf_smut_df = pd.DataFrame(data_leaf_smut)

all_data = pd.concat([bacterial_leaf_blight_df, brown_spot_df, leaf_smut_df])

all_data

from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import train_test_split
from sklearn import metrics

input = np.array(all_data)[: , :784]
output = np.array(all_data)[: , 784]

input_train, input_test, output_train, output_test = train_test_split(input, output, test_size=0.2, random_state=42)
model = DecisionTreeClassifier(max_depth=1)  
clf=model.fit(input_train, output_train)
output_pred=clf.predict(input_test)
print("accuracy: ", metrics.accuracy_score(output_test,output_pred))

read_photo1 = img.open(test).convert("L")                                                                                                                                          
resize_photo1 = read_photo1.resize((28,28))  
normalize_photo1 = np.array(resize_photo1)/255.0  
convert_photo1=normalize_photo1.flatten()   
convert_photo1=np.reshape(convert_photo1, (1,-1))

print(clf.predict(convert_photo1))

if clf.predict(convert_photo1) == 0:
    text = "bacterial leaf blight.."
    print(text)
elif clf.predict(convert_photo1) == 1:
    text = "brown spot.."
    print(text)
elif clf.predict(convert_photo1) == 2:
    text = "leaf smut.."
    print(text)
    

import cv2
photo=cv2.imread(test)
cv2.putText(photo, text, (10,230), cv2.FONT_HERSHEY_SIMPLEX, 1, (255,100,155),2)  
cv2.imshow("tahmin", photo)
cv2.waitKey(5)   
cv2.destroyAllWindows

from sklearn.metrics import confusion_matrix
cm=confusion_matrix(output_test, output_pred)
print(cm)
