==> week3/19-gradient_booosting.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import confusion_matrix
from xgboost import XGBClassifier

data=pd.read_csv("kanser.csv")

data.head()

data.drop(["id", "Unnamed: 32"], axis=1, inplace=True)

data.diagnosis=[1 if each =="M" else 0 for each in data.diagnosis]

data

y=data.diagnosis.values
X=data.drop(["diagnosis"],axis=1)

Xtrain,Xtest,ytrain,ytest = train_test_split(X,y,test_size=0.2,random_state=42)

sc=StandardScaler()
Xtrain=sc.fit_transform(Xtrain)
Xtest=sc.transform(Xtest)   #we apply scaler only for xtrain and xtest

xgb=XGBClassifier(eval_metrics="logloss",
                  n_estimate=100,
                  max_depth=3,
                  subsample=1,
                  colsample_bytree=1,
                  random_state=42)

xgb.fit(Xtrain,ytrain)

ypred=xgb.predict(Xtest)
accuracy=xgb.score(Xtest,ytest)  #!!!
print("accuracy:",accuracy)

cm=confusion_matrix(ytest,ypred)
cm

from sklearn.ensemble import GradientBoostingClassifier
gb=GradientBoostingClassifier(n_estimators=100)

gb.fit(Xtrain,ytrain)

ypred=gb.predict(Xtest)
accuracy=gb.score(Xtest,ytest)
print("Accuracy: ",accuracy)
