==> week3/22-random_search.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split,RandomizedSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import confusion_matrix

data=pd.read_csv("kanser.csv")

data.drop(["id", "Unnamed: 32"], axis=1, inplace=True)

data.diagnosis=[1 if each =="M" else 0 for each in data.diagnosis]

y=data.diagnosis.values
X=data.drop(["diagnosis"],axis=1)

Xtrain,Xtest,ytrain,ytest = train_test_split(X,y,test_size=0.2,random_state=42)

sc=StandardScaler()
Xtrain=sc.fit_transform(Xtrain)
Xtest=sc.transform(Xtest)  

from sklearn.ensemble import RandomForestClassifier
model=RandomForestClassifier(random_state=42)

param_random={
    "n_estimators": [25, 50, 75, 100, 150, 200],
    "max_depth": [5, 8, 12, 15, 20, 22]
}

rm=RandomizedSearchCV(model, param_random, cv=5, n_iter=10, scoring="accuracy", n_jobs=-1, random_state=42)   #10 ihtimal için çekim yapacak cross validation da olduğu için totalde 10x5 50 defa model eğitilecek
rm.fit(Xtrain, ytrain)

print("The best parameters: ",rm.best_params_)
print("The best score: ",rm.best_score_)

best_test=rm.best_estimator_
accuracy=best_test.score(Xtest,ytest)

print("The best score for test data: ", accuracy)
